+++ NEWS.md
@@ -1,4 +1,7 @@
-# rxode2 (development version)
+# rxode2 development version
+
+- A bug was fixed so that the `zeroRe()` function works with correlated omega
+  values.
 
 # rxode2 2.0.12